requirements.txt
schedule==1.1.0
requests==2.27.1
psycopg2==2.9.3
python-dotenv==0.20.0
flask==1.1.2